Major cleanup of main() and clicker
Major things is a clean up of the project: code was too messy to maintain and add features to, made a helper object class that holds most functionality.

That helper object is then inherited from to my actual "doer" classes: the fisher (which does the fishing in version 1) and now also added the clicker (which finds, buys and sells potions).